backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG Backend with FastAPI, Qdrant, and vLLM"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.3.0",
    "vllm>=0.2.7",
    "python-multipart>=0.0.6",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.2",
    "lxml>=5.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
